Allow unit tests to bypass mocks if apiKey is set
With the changes in this commit, the unit tests will allow the MISP
service to make real requests to the MISP server if the apiKey is
pre-configured in the dxlmispservice.config file. If the apiKey is not
pre-configured, the tests will continue to use mocks. Assertions against
response data were also expanded to validate more of the data that the
samples use.