Update attribute UUID handling in basic update event example test
Previously, the request mock validation in the basic update event
example test expected the UUID for the attribute that the sample creates
to be assigned by the MISP server. Starting with PyMISP v2.4.90.1,
however, the PyMISP library generates a UUID for the attribute on the
client, which caused some of the test assertions to fail.

This commit adjusts the mocks and assertions in the update event
test to use the client-generated UUID in the request, if present, or
continue to ignore it, if not present. These changes allow the tests to
pass on both the older and newer PyMISP versions.