ZeroMQ socket handling cleanup to avoid shutdown hang
Previously, unit tests intermittently failed due to a hang when running
in Travis CI - frequently on Python 3.4. The hang was due to the
poll() for ZeroMQ MISP notifications being done on a background thread
not being interrupted and allowing the thread to be terminated as the
MISP DXL service was shutting down.

In this commit, a separate PULL socket associated with an internal PUSH
socket is registered with the ZeroMQ poller. During DXL service
shutdown, a message is sent to the PUSH socket, allowing activity on the
PULL socket to interrupt the blocking ZeroMQ poll() and stop the poll
thread.

This commit also includes some additional logic for freeing the ZeroMQ
context and sockets during shutdown.